+++ Classes/Table/Football/ComparatorH2H.php
@@ -3,6 +3,7 @@
 namespace System25\T3sports\Table\Football;
 
 use System25\T3sports\Table\Util;
+use System25\T3sports\Table\IComparator;
 
 /***************************************************************
  *  Copyright notice